* implemented token blacklisting for invalid pans tokens by using the APNS extended protocol
Invalid APNS tokens i.e. sandbox tokens will cause Apple to drop the
APNS SSL connection. By using the extended protocol we can provide an
ID with each message and Apple will send a binary notice prior closing
the connection with the token causing the problem and a reason code.
Tokens are stored in a "dict buffer" for later lookup. If a token
causes a connection drop, the token will be added to a global black
list (per appID). Next time the token will be filtered out. This will
at least provide a level of "self-healing" - Next step would be to
expose the black list via XML-RPC so that the client can fetch it (
like the feedback service data)